Extract sidebar panels from App
